requirements.txt
python-multipart
uvicorn[standard]
fastapi
openai
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pymongo
pydantic